refactor(client): renames files and folders
