+++ client/src/pages/signup/index.ts
@@ -1,3 +1,3 @@
-import SignUpContainer from 'pages/signUp/signup.container';
+import SignUpContainer from 'pages/signUp/signUp.container';
 
 export const SignUp = SignUpContainer;
